requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
alembic==1.15.1
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
distro==1.9.0
dnspython==2.7.0
docx==0.2.4
email_validator==2.2.0
exceptiongroup==1.2.2
faiss-cpu==1.10.0
fastapi==0.115.11
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
joblib==1.4.2
lxml==5.3.1
Mako==1.3.9
MarkupSafe==3.0.2
multidict==6.1.0
numpy==2.2.4
openai==1.66.3
oso==0.27.3
packaging==24.2
passlib==1.7.4
pdfminer==20191125
pdfminer.six==20240706
pillow==11.1.0
propcache==0.3.0
psycopg2-binary==2.9.10
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
PyJWT==2.10.1
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.20
scikit-learn==1.6.1
scipy==1.15.2
sniffio==1.3.1
SQLAlchemy==2.0.39
starlette==0.46.1
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.12.2
uvicorn==0.34.0
yarl==1.18.3
